Modified the SQL schema to include site information
I decided to go ahead and put the site information here because this
is really the code that needs the information on when TCCON (or other)
sites where in different locations. I can add a JSON API to the rocket
app that delivers this information as a JSON that ginput can parse
and/or cache.