Have a CLI command to get the JSON site info working
Note: to get VS Code to recognize the tokio::main macro,
I had to change the settings for this workspace, so we need
the .vscode/settings.json file.

The CLI will also be asynchronous because the ORM needs to be
async for Rocket, and it's just simpler to do this (enabling the
use of `await`) than to always block on ORM calls. This may also
open up multithreading for the CLI if it seems advantageous.