Started working on unstyled pages for the Rocket app
The main thing I did was start on the nav bar. The tricky part
was making the URLs robust - Rocket provides a `uri!` macro that
can look up the proper URL given a function and any parameters,
but this cannot be used directly in templates.

One option would have been to make a struct that holds the nav bar
URLs as fields and always provide this as a field of each page's
context object, but it has to be a field, and would need to be the
same field name. I tried creating a trait with a function to return
this, but the templates couldn't call the function.

The best solution turned out to be implementing the `Function` trait
from Tera on this struct. Then I could register it as a global function
and call that function in the nav bar template. Thus, the different
pages don't need to worry about providing the nav bar URLs in their
context objects.