Migrated to configurable met products
This commit should remove the old dependence on the assumed GEOS
file organization, and switch completely to using the met files in the config.
This required implementing the "default options" table described in issue #9.
Having implemented that, I now think it makes more sense to be part of the
config. That rationale is that the config should be where things only changed
by humans go, and the database is where things mostly changed by the program
go.