Changed migration to satisfy MySQL on loaner Mac
Not sure if this is because I used a more recent MySQL version,
because I used `cargo sqlx migrate run` for the first time, or
because I hadn't tried running the migration in this form before, but
on the loaner laptop, the MySQL version did not like variable length
fields with unique constraints and some other things. This commit may
be revertable on another machine.

I did find out that you can make MySQL calculate SHA256 hashes on
text automatically (see https://stackoverflow.com/q/53617727), which
means I can have it automatically handle generating a fixed length
hash of the met file paths to enforce uniqueness on.