`Job` queries better at avoiding race conditions
Added some new functionality to `Job` in the core ORM to handle
getting the next job to run and setting its state to `running`
without the possibility of a race condition between two threads,
where two threads end up starting the same job because the query
to get the highest priority job happened in both threads before
either thread set the state to `running`.

I've changed how the `job` module returns errors to mostly use
a `JobError` type rather than the `anyhow` error. This made it
easier to check if a job query deadlocked, which can happen if
the first thread to query for the next job is the second one
to try setting its state to `running`.