Fixed ginput jobs blocking the scheduler thread
The solution in the end was to replace the `Command` instance running
`ginput` with the `tokio` version instead of the `std` one. It seems
that the problem was, because the `std` one was running for a long time
without hitting an `await`, that thread wasn't letting other threads
try to proceed. However, that can't be the whole reason, because the
jobs to regenerate the LUTs also (and currently still do) use the `std`
`Command` as well. The main difference I saw with the job thread is
that it needed to hold a database connection. This commit has that
holding a pool instead, and only creating short-lived connections as
needed, but that may not be necessary. I'm going to try reverting that
in the next commit; if the next commit doesn't change the pools back
to connections, then limiting the lifetime of connections held to the
database was important as well.