Input file failure emails and moving files implemented
We now move input files into their success/failure directories and
use the configured email backend to email the user if there is a
formatting issue. Because moving the files can fail, and because it
would be a problem if a file failed to get deleted from the input
directory, the input file handling logic has been modified to
(theoretically) handle that case without spamming the admin emails.