WIP towards standard site jobs
Added skeleton of service manager. Fixed some bugs in the met code
(especially for rescanning for met files). Reworked how the standard
site jobs module in the ORM scans for jobs, should be more robust
but is incomplete. In particular the code to make the jobs themselves
is unfinished and will change in the coming commits.