Adding jobs for standard sites implemented
I had to add a new `MapFmt` variant - `TxtAndNc` - that represents a
request to generate both text and netCDF files. I did this so that the
standard site jobs just generate all the files they could possibly need
(generating the `.map` files is quick) and we can cherry pick the ones
we want while making the standard site tarballs.

I deliberately named the variance `TxtAndNc` rather than `All` to remind
me to update `ginput` in the future if we happen to add more `.map`
formats.