Fixed standard site job submission to mark days as MissingMet
Now when submitting standard site jobs, if one of the days we try to
run is missing meteorology, its status in the table is changed to
`MissingMet` and a job will not be submitted for it. The next table
update will recheck this day and reset it to `JobNeeded` if the
meteorology has been added.

Also, the functions to update the standard site jobs table now by
default only consider days after the first day with valid meteorology,
rather than the start of the site's operation.