Update to jobs: output directory and completion emails
Jobs will now run in a `.running` subdirectory of their `save_dir` - I did
this rather than letting the run directory be fully configuable because the
easiest way to move a directory in Rust is just to rename it, but renaming
doesn't work across filesystems. By making the run directory a subdirectory
of the save directory, that is unlikely to happen.

I also had not implemented the "job complete" and "job errored" emails, that
is now done.