Fix bug in `DateIterator` causing met download errors.
Previous, if given a date range where the start and end dates were the same,
the `DateIterator` would incorrectly still yield the first date in the range.
This was causing the service met download to try to download data too early,
because it would get that the first date required was today, and set the end
date to today. This should have produced a no-date iterator, but instead
yielded today. Since GEOS data is not available that quickly, we get an error.