Fixed bugs with setting site to non-op
Output files weren't being cleared and the job rows weren't being flagged.
It was also cleaner to remove the "nonop" state and just have a site's abscence
from the jobs table on the date indicate that, that way the normal logic to
regenerate it if the site is added back later works fine.