Met downloading should be smarter about latency.
In testing, I was getting errors when the UTC time advanced by a day
(causing the downloader to try to download today's GEOS data) but not
all the GEOS files were ready yet. It should now differentiate a
critical error from one caused by the files just not being ready yet.

Also, it was downloading multiple copies of the 3D met files, because
it would find e.g. the 0000 UTC file, but none of the other the first
time it was called, then the 0000 to 0600 UTC files the second time,
etc. I've modified the `WgetDownloader` to use timestamping, which
should avoid that.