Added CLI commands to generate special runs of standard site priors
The motivation for this was that I will need to make GEOS IT priors for
a number of test periods in 2018 and 2022/23. I don't want to have to
manually submit jobs for all the standard sites, plus I wanted to use
the existing code to make the tarballs.

This added two new commands to the `site-jobs` CLI, one to submit jobs
to generate priors using arbitrary ginput and met types and one to tar
those files when the jobs are done.

This did necessitate a change to the tarball code. Previously it assumed
that jobs being made into standard site tarballs would only have one day,
so it could just use the start date of that job for the tarball name
and only make one tarball per site. Because I allow the special runs to
have >1 day per job, this no longer works. The tarball-making code on
`StdOutputStructure` is more robust now, and will loop through the dates
for a given job. However, `make_standard_site_tarballs` in
`orm::stdsitejobs` still assumes that only one tarball is made per job,
as it needs to insert that tarball name in the `Jobs` table output file
field. This should be fine, because the automatic standard site jobs
will still follow that convention.