User requests can now be allowed to use non-default reanalysis
By default, user requested runs will only be allowed to use the defaults
specified in the config for each date. I've added the ability to configure
alternates. This would require new config sections, e.g.:

```
[requests.allowed_mets.co_reprocessing]
met_key = 'geosit_chm'
ginput_key = 'v1.2.0'
end_date = '2024-04-01'
```

The user would include the line `reanalysis=co_reprocessing` in their
input file to tell the automation to use this combination of the met and
ginput. In this case, if they requested dates on or after 1 Apr 2024,
the request will be rejected. This was simpler than truncating the
request. A `start_date` can also be included.