Login is functional, needs some quality-of-life work.
The log out function exists, but does not have a sidebar entry yet
because it should only show up when a user is logged in. So I will
need to figure out how to use auth session to get a user if logged
in. That will then need added to all view functions and the base
context, but will also mean I can put the user name in the header
bar when logged in.