Service tests now use testcontainers
Moving the helper functions for dealing with testcontainers into the ORM was
trickier that I expected. Because integration tests don't include code gated
behind a `#[cfg(test)]`, it wasn't as straightforward as I hoped to make
testcontainers a conditional dependency. The only way I found to do it was
to have the actual code that uses it gated behind a "container-tests" feature,
and to give the init function for a containerized test DB a panicking version
when "container-tests" is not enabled. This means that tests will need to be
run with `--feature container-tests` to be able to use containers now.

If we go fully to containerized testing, with no capability to use a persistent
DB, then I might be able to use a `compile_error!` instead of a `panic!` to ensure
that tests are always run with that feature. But this is fine for now.