Checkpointing axum and askama update
`askama` went through a bigger change, with a fork being merged back
into the original repo. I have not run the web server yet, but on the
compilation side, I had to deal with the deprecation of `axum-askama`.
Now instead of template implementing `axum`'s `IntoResponse` directly, I
have to call `.render()` on the templates.