First pass at refactoring to reduce request check duplication
I did a lot of refactoring on the code that handles request checks. The
input file module still has checks that the input file has all of the
necessary fields, but checks for the validity of the request are now in
the `orm::jobs::requests` module so they can be reused between the input
files and the web interface. There's still some duplication among the
different `orm::jobs::Job::add_*` functions, but I haven't decided
if/how to simplify that yet. I also really need to run the input file
integration tests to verify all the checks still work and write new ones
for the API requests.