Completed first draft of the website API
All the essential components for the API are in place and work locally.
There are still several elements needed for the web service to be fully
functional, however:

1. A page for users to manage their API tokens.
2. Modify the job download page to use streaming endpoints like the API
3. Allow user "groups" permission to download each others' jobs
4. Add a database migration to assign jobs to users by ID, not email.
5. Include the server address in the configuration and ensure it is
displayed correctly in all relevant locations, e.g. in the API docs
6. Include the location of the static files in the configuration.

The database migration itself should not be too bad - just a matter of
adding the user ID as a nullable foreign key. I should edit things so
that the emails can be a list for additional flexibility, but that isn't
essential. The bigger challenge is going to be adding all of the current
users requesting jobs to the database. Then I have two approaches to
migrate users:

1. Send an email to past users to confirm who is still using the service
with a google form. Then I can cross reference those emails against the
FTP list, create accounts for them, and have them reset their passwords.
2. Create a page where users can enter their email and the backend
checks it against the FTP list (and ensures it hasn't already been
used).

I prefer 1, so that I have control over who accesses the priors.