Checkpointing met refactor once more
I have the met download code modified to at least compile with the new
config, but I think it may need to be revisited. My concern is that
it may be set up to only download one set of met files automatically
per day, and that will need to be made more flexible.

I'm getting ready to modify the `Jobs` table to use just a processing
key instead of separate ginput and met keys.