Checkpointing first effort to update standard site jobs to parallel processing streams
I'm not happy with how complex the logic is looking to be to determine
the start and end dates for updating the site table. Since that depends
on both the available met and the configured processing, it seems like
it will be better to iterate over the automatic processing
configurations higher up, but that will mean other things need to
specify a processing config.

First I want to make the jobs' processing configuration clearer about
whether it is a default or undefined by switching to a custom enum
instead of an `Option`.